final
all needed outputs

+++ kristofidies/City.cs
@@ -2,8 +2,8 @@
 {
     public class City
     {
-        public double X { get; set; }
-        public double Y { get; set; }
+        public int X { get; set; }
+        public int Y { get; set; }
         public int Number { get; set; }
     }
 }
